change userId from int to long

+++ app/src/main/java/com/sitexa/android/community/internal/di/modules/UserModule.java
@@ -34,12 +34,12 @@ import dagger.Provides;
 @Module
 public class UserModule {
 
-    private int userId = -1;
+    private long userId = -1;
 
     public UserModule() {
     }
 
-    public UserModule(int userId) {
+    public UserModule(long userId) {
         this.userId = userId;
     }
