+++ domain/src/main/java/com/sitexa/android/domain/repository/UserRepository.java
@@ -35,5 +35,5 @@ public interface UserRepository {
      *
      * @param userId The user id used to retrieve user data.
      */
-    Observable<User> user(final int userId);
+    Observable<User> user(final long userId);
 }
